add: avoid firebase json upload on s3

+++ frappe_s3_attachment/controller.py
@@ -193,6 +193,9 @@ def file_upload_to_s3(doc, method):
     """
     check and upload files to s3. the path check and
     """
+    if getattr(doc, "custom_skip_s3_upload", 0):
+        frappe.logger().info(f"Skipping S3 upload for {doc.name} due to skip_s3_upload flag.")
+        return
     s3_upload = S3Operations()
     path = doc.file_url
     site_path = frappe.utils.get_site_path()
@@ -324,6 +327,9 @@ def migrate_existing_files():
 
 def delete_from_cloud(doc, method):
     """Delete file from s3"""
+    if getattr(doc, "custom_skip_s3_upload", 0):
+        frappe.logger().info(f"Skipping S3 remove for {doc.name} due to skip_s3_upload flag.")
+        return
     s3 = S3Operations()
     s3.delete_from_s3(doc.content_hash)
